+++ src/datetime.cpp
@@ -1,9 +1,49 @@
+// NOLINTNEXTLINE(modernize-deprecated-headers): no thread-safe version in STL
 #include "databento/datetime.hpp"
 
-#include <iomanip>
-#include <sstream>
+#include <time.h>  // gmtime_r or gmtime_s
+
+#include <array>
+#include <chrono>
+#include <ctime>    // localtime, strftime, tm
+#include <iomanip>  // setw
+#include <sstream>  // ostringstream
+
+#include "databento/constants.hpp"  // kUndefTimestamp
 
 namespace databento {
+std::string ToIso8601(UnixNanos unix_nanos) {
+  if (unix_nanos.time_since_epoch().count() == kUndefTimestamp) {
+    return "UNDEF_TIMESTAMP";
+  }
+  std::array<char, 80> buf{};
+  const auto time =
+      static_cast<std::time_t>(std::chrono::duration_cast<std::chrono::seconds>(
+                                   unix_nanos.time_since_epoch())
+                                   .count());
+  std::tm tm = {};
+#ifdef _WIN32
+  if (::gmtime_s(&tm, &time) != 0) {
+    // Fallback on printing nanos
+    return ToString(unix_nanos);
+  }
+#else
+  if (::gmtime_r(&time, &tm) == nullptr) {
+    // Fallback on printing nanos
+    return ToString(unix_nanos);
+  }
+#endif
+  const auto nanos = std::chrono::nanoseconds{
+      unix_nanos.time_since_epoch() %
+      std::chrono::nanoseconds{std::chrono::seconds{1}}};
+  const size_t count =
+      std::strftime(buf.data(), sizeof(buf), "%Y-%m-%dT%H:%M:%S", &tm);
+  std::ostringstream time_ss;
+  time_ss.write(buf.data(), static_cast<std::streamsize>(count));
+  time_ss << '.' << std::setw(9) << std::setfill('0') << nanos.count() << 'Z';
+  return time_ss.str();
+}
+
 std::string ToString(UnixNanos unix_nanos) {
   return std::to_string(unix_nanos.time_since_epoch().count());
 }